Produce a single binary for integration tests
If we have multiple files here, they end up in different test binaries.
This means that we have to link everything separately.
And linking so much is slow (and likely single-threaded).

Note: modules need to be moved into subdirectory to not be linked separately.

See: https://endler.dev/2020/rust-compile-times/#combine-all-integration-tests-in-a-single-binary

Co-authored-by: Restioson <restiosondev@gmail.com>