Roll-out MakerOffers into projection and extend sse api
We opted for a `HashMap<ContractSymbol,...>` to be able to handle multiple contract symbols and subscribe to them individually via the sse feed. This is more efficient as otherwise our frontend would constantly reload if _any_ symbol received a new quote, even if we do not care at the time.