Enable discrete payout generation based on quanto curve
In order to generate a `quanto::Curve` model you have to provide the
same arguments needed for the original inverse payout curve, plus a
`multiplier`.

The API `model::Payouts::new_quanto` could eventually be combined with
`model::Payouts::new_inverse`. Passing in the `ContractSymbol` would
then allow branching between inverse and quanto. The `ContractSymbol`
would also imply the `multiplier` to be used in the case of quanto
curves.