Log `peer_id` in context of finishing a protocol
This helps when debugging problems, because we don't have to first associate the `order_id` with the `peer_id`.
Note that this has to be logged as option for as long as we have backwards compatibility to the old connection.

We could opt for moving this log into the protocol, but that results in a mess because we then have to duplicate this log multiple cases to catch all cases.