Remodel bankruptcy price for quanto curve
It did not actually make sense to combine the previously called
`LiquidationThresholds` into a single type, because each instance of
the type was modelling these values for parties with the _same_
initial margin (i.e. the same leverage). This is the reason why we had
to instantiate it twice (once for long and once for short) with
different initial margins.

Furthermore, the field `bankruptcy_price` was incorrectly named. What
we were calling "liquidation thresholds" were actually the bankruptcy
prices! The field was actually the `bankruptcy_price_shift`: the
absolute change from the initial price that would cause a position
with a certain initial margin to reach its bankruptcy price. This is
now correctly modelled with a free function.