Handle overlapping liquidation intervals for quanto curve
This is an edge case that most likely only appears in proptests. We
already saw this in CI:
https://github.com/itchysats/itchysats/runs/7685221192?check_suite_focus=true.
But it's hard to avoid, so we either need to handle it or explicitly
`reject` the errors like we do with `{Long,Short}OwesTooMuch`.

I chose to handle it because it's less code. I chose to increase
the lower bound of the short liquidation threshold rather than to
decrease the upper bound of the long liquidation threshold to avoid
possibly getting a -1 as a result, which would panic.