refactor(sqlite-db): Use SqliteConnection instead of Transaction
When we started to work with the `sqlx::Transaction` type to ensure
that certain database transactions were either fully committed or
rolled back, the type started leaking to lower level calls that should
not need to care about whether the caller is using an
`sqlx::Transaction` or a regular `sqlx::PoolConnection`.

By using the `sqlx::SqliteConnection` type in function type signatures
we are able to abstract over both `sqlx::Transaction`s and
`sqlx::PoolConnection`s. Because of this, we can greatly reduce the
boilerplate in `sqlite-db` tests.