feat(feed): Remove contract symbol from offer feed path
It should make it easier for us to implement and consume this feed if
the client simply filters the events they're interested in.

The filtering could eventually be achieved through a query parameter
on the feed URL, but currently we just filter on the client side.

This change temporarily introduces a regression: If a client
reconnects to the feed (e.g. by refreshing the UI) they will only get
the last BTCUSD or ETHUSD offer.

Co-authored-by: Richard Holzeis <richard@coblox.tech>