logs(endpoint): Do not emit error event if opening a substream fails
Failing to open a substream will almost always result in the caller
handling the error, which will almost certainly be logged. Therefore,
there is no need for emitting a redundant error event via the
`tracing::instrument` macro.

Furthermore, this will allow callers to decide which errors should be
logged and which can be silenced.