feat(network): Add /itchysats/offer/2.0.0
The `/itchysats/offer/2.0.0` protocol offers proper support for
sending and receiving offers with different contract symbols. The
design of the message exchanged between maker and taker has changed to
make this easier. These are the key differences:

- Before, we only allowed the maker to send a pair of offers that were
supposed to reference the same contract symbol, which was sufficient
for BTCUSD. Now, we give freedom to the maker to send a list of
offers. The taker must combine the new offers with the old ones to
derive a full view of the latest offers.

- Before, parts of the message were optional to allow the maker to
"disable" certain offers and even rollovers. This feature is not
needed since both the maker has other mechanisms in place to reject
orders and rollovers. Now, all fields are mandatory (sending an empty
list is technically possible, but it doesn't mean anything for the
taker).

The maker supports versions 1.0.0 and 2.0.0 side by side. We maintain
version 1.0.0 for the maker in a `deprecated` module, but remove a lot
of the code that is no longer necessary (the taker side and the
tests). We have adapted what remains so that it is easier to integrate
it in `maker` together with the new protocol.

Because the maker is the dialer in the offer protocol, it means that
currently all their takers will get doubled offers (one per version)
and will ignore the ones that correspond to the protocol they don't
support. To prevent us from generating false positive warnings on the
maker side, we filter out the `yamux::NegotiationError::Failed` error.

Additionally:

- Rework the way the offers are modelled in different components
across the entire codebase.
- The `projection::Actor` now takes a `Role` parameter, so that we can
do away with the awkward `Origin` type in the new offer protocol.
- Simplify `model::Offer`, which is not actually used anywhere
where backwards-compatibility matters.
- Calculate `oracle_event_id` based on `settlement_interval` when
constructing `Offer`. As opposed to passing two arguments that could
be out of sync, like we used to do.