feat: Introduce contract `lot_size`
Lot Size (in number of contracts) is the minimum trading unit of a contract. Order quantity need to be exact multiples of the contract’s Lot Size. Order quantity not in multiples of Lot Size will be rejected.

For example, a contract with Lot Size of 100 would allow order quantities of 100, 200, 500, or 1000 contracts but would reject quantities of 1, 5, 10, 35, 125, or 250 contracts.

Resolves #2660

Signed-off-by: Philipp Hoenisch <philipp@coblox.tech>