refactor(model): Model payout as a bitcoin::Amount
We solve a TODO that is actually in `daemon::projection`, but which
stems from wrongly representing payouts as `bitcoin::SignedAmount`s. Payouts
are by definition non-negative.

Even though this modifies the public API of `projection::Cfd`, I don't
think it is a breaking change because any client that expects this to
be a `bitcoin::SignedAmount` will be able to handle a
`bitcoin::Amount`.