refactor(model): Rename calculate_profit_at_price
The function is primarily calculating the payout, so the name
`calculate_payout_at_price` seems more appropriate. As the very last
step we are calling a different function which calculates the profit.

It would be nicer if we didn't have to return 3 orthogonal things from
the same function, but it is convenient because otherwise we would
have to calculate the margin again on the outside in order to get the
PNL.