chore(model): Make Price conversion into u64 infallible
This can only fail if the `Decimal` is too big to fit into `u64` or if
it is negative.

It cannot be negative because a negative price would be a bug. If the
price is too big that it cannot fit into a u64, it would be a bug
anyway and we don't need to protect against that.