refactor(model): Separate profit calculation from payout curve
Calculating the profit is independent of the payout curve used.

We also modify `model::cfd::calculate_payout_at_price` so that it only
returns the payout, instead of a tuple including PNL values. It's
generally better if we can write several functions that compose well,
rather than one function that does everything at once. In this case it
makes sense to split this functionality because we were already using
the `calculate_profit` function separately too.