feat: Calculate funding fee depending on the contract symbol
The funding fee calculation uses the initial margin which depends on
the contract symbol. Before, this value was hard-coded to
`ContractSymbol::BtcUsd`, but now it is dynamic.

The `FundingFee` type is constructed: when the CFD is created; every
time the CFD is rolled over; and when building CFD projections for the
daemon feeds.