feat: Calculate liquidation prices depending on contract symbol
Before, we were using a formula that was specific to inverse payout
curves (for the BTCUSD contract symbol). Now, we choose the formula
dynamically based on the contract symbol.

Interestingly, calculating the liquidation prices for ETHUSD involves
the number of contracts, unlike BTCUSD. This leads us to having to
choose an arbitrary value (between `min_quantity` and `max_quantity`)
when calculating the liquidation price for the offers in the
`projection` feed.