feat: Support BETH Olivia event IDs
With this change we will choose oracle event IDs based on the contract
symbol.

This makes the serialisation of the `BitMexPriceEventId` dynamic over
the new `index: IndexPrice` field, derived from the
`model::ContractSymbol` type.

Interestingly, even though we have to expand the `sqlite-db` types
that are used for loading from and saving to the database, these
changes aren't breaking. The original `BXBT` events are still
represented the same way.