Log all events in one place
Stop logging events in the model all over the place and created on dedicated way to log all events.
This helps with following the history of a cfd and enhance the logs with further information in a structured way (e.g. log the `peer_id` for all events issued).

Given that we want to log most events on `info`, but some on `warn` and `error` we distinguish by event kind.