feat/fix (quanto): Use effective leverage
For calculating the liquidation price we have to take the updated fee margin (reduced or increased due to fee account) into account. This means, the liquidation price will change in the lifetime of a position, i.e. on every rollover.

Note: this commit also adds . A proptest which was failing before but is fixed with this commit.