fix(olivia): Use all fields in BitMexPriceEventId Hash impl
If we do not include the `index: IndexPrice` field in the derived
`Hash` implementation of `BitMexPriceEventId` we will not be able to
differentiate between `BXBT` and `BETH` events when using
`BitMexPriceEventId`s as keys in `HashMap`s and `HashSet`s.

Similarly, if we do not include the `digits: usize` field we will not
be able to differentiate between events with different values for this
field. This is currently irrelevant as Olivia only supports price events
with 20 digits, but there is no need to maintain this wrong behaviour.