fix: Do not overwrite inverse payout curve liquidation bound
This reverts 2298059164d39ea744b708e6f54dcee46ab96037. Whilst this
fixed https://github.com/itchysats/itchysats/issues/2571, it also
meant that takers without this fix would not be able to perform
contract setup and rollover with a maker with this fix.

We will very likely introduce an equivalent fix in the future, but it
cannot be safely introduced without changing `/itchysats/order/1.0.0`
or introducing a new version altogether.