chore(projection): Model liquidation price as Decimal
We gain nothing from using the `model::Price` type, which imposes a
questionable on the values that it can represent: the `model::Price`
type doesn't allow the value of zero.

This change allows us to remove the post-processing of an edge case
when calculating the long liquidation price for quanto payout curves.