fix(projection): Use actual liquidation price for open CFDs
We can deduce the liquidation prices for open CFDs by looking at the
stored CETs, as soon as we have them.

This is only possible for open CFDs, because they are the only ones
that have access to this information (unlike offers, closed CFDs and
failed CFDs). Having correct liquidation prices also matters a lot
more for open CFDs, as the user can actually use this information to
make decisions.

We've had to modify a `projection` test which ensures that a
`projection::Cfd` didn't change meaningfully after transitioning from
open to closed. We expect the liquidation price to always differ now,
because it's calculated differently depending on whether the CFD is
open or not. This is obviously not ideal and it should be fixed,
probably by recording the liquidation price explicitly when archiving
CFDs into the `closed_cfds` and `failed_cfds` database tables.

We've also had to modify the expected liquidation prices in some
daemon tests, given that they are now calculated differently.