feat(offer): Forbid sending empty offers via /itchysats/offer/1.0.0
By leveraging the `nonempty::NonEmpty` type we can use the type system
to ensure that we will never send empty offers to takers via
`/itchysats/offer/1.0.0`.

The problem with sending a `MakerOffers` with `None` values for `long`
and/or `short` is that takers will interpret this as an instruction to
delete their local offer state. That is something that we no longer
want to trigger from the maker's perspective.

In particular, this patch fixes a bug where any `taker:<=0.5.5` was
deleting its local offer state every time the maker was trying to send
ETHUSD offers via the `/itchysats/offer/2.0.0` protocol. This was
precisely because we were filtering out all ETHUSD offers derived from
the `OfferParams`, resulting in an empty list of offers.