Fix counting already connected peers
In some cases we are not notified that a connection was dropped by the underlying network stack.
Thus, it can happen that we just replace a `peer_info` in the hashmap without removing it first.
Incrementing the gauge on every insert can thus lead to counting connections for one peer multiple times.
We now ensure that we only increment if the `peer_info` was not recorded already.