Supply a config file with otel-tests tracing directives
Removes the need for the crate to know about other crates that are set further
down the dependency chain and allows for extracting this generic crate out of
the itchysats workspace.

New `otel-tests.toml` file needs to be located in the directory of the crate
using `otel-tests` macro for tests.

In case the file is not present or it is malformed, an otel-test will fail to compile.