fix(offer): Tweak dialing error handling
Before, we just suppressed any `xtra_libp2p::Error::NegotiationFailed`
resulted from trying to open a substream to perform an
`/itchysats/offer` protocol.

Now that the `Endpoint` gives us more context, we can improve the logs
we emit:

- If the `Endpoint` knows that the peer doesn't support the protocol,
we should suppress the error and move on.
- If the protocol fails during the negotiation phase, we log the
failure on debug. This could happen for peers that did not yet tell us
about the listen protocols they support. If we saw this log very
frequently we could start investigating a possible bug.
- Any other error still warrants a warning.