chore(rollover): Remove rollover Version enum
We now use separate protocol actors to support different versions of
protocols, so it makes sense to remove the `rollover::Version` enum
from that perspective.

We also only need to support the latest variant in that enum (because
all other variants correspond to the legacy network stack), which
means we can remove the enum altogether.