feat(payouts): Allow configuration of inverse max price
With the new configuration enum `InverseMaxPrice` we can optionally
fix a bug[1] with the short liquidation events of the inverse payout
curve.

The reason why we don't unconditionally fix the bug is to maintain
backwards compatibility for all network protocols that depend on this
API. In fact, this patch sets the value to
`InverseMaxPrice::DoubleOfInitial` in all cases for now. A future
patch that introduces new versions of different network protocols will
introduce the use of the other variant.

[1]: https://github.com/itchysats/itchysats/issues/2571.