chore: Replace chashmap with dashmap
The `chashmap` dependency has lead to dependabot alerting[1] about a
vulnerability in a transitive dependency. Looking at `chashmap`'s
repository it looks to me like it's not very well maintained.

On the other hand, `dashmap` looks great and should be a good
replacement for `chashmap` and `chashmap-async`. Even though `dashmap`
does not yet support async locks[2], the maintainer states that
synchronous locks "are also fine to use in async as per tokio
docs"[3].

[1]: https://github.com/itchysats/itchysats/security/dependabot/35.
[2]: https://github.com/xacrimon/dashmap/issues/150.
[3]: https://github.com/xacrimon/dashmap/issues/150#issuecomment-1030858774.