fix(monitor): Handle early states and RolloverRejected differently
When re-hydrating the `monitor::Cfd` we do not need to do anything to
the aggregate when handling `ContractSetupStarted`,
`ContractSetupFailed`, `OfferRejected`. These states can only occur
right after the creation of the CFD, so there is nothing that needs to
be enabled or disabled in terms of monitoring if contract setup didn't
succeed.

For `RolloverRejected`, we should definitely not disable things like
monitoring for the commit transaction finality. That could see us
failing to rollover and then not checking if the other party has
published the current commit transaction!