fix(monitor): Separate monitor logic from transaction data
When we worked on b4225d5d35e27caef749d5c8ebf4e0f28c55d649 we
inadvertently introduced a bug[1]. The gist of it is that, after a
restart following a commit transaction being confirmed, we were not
monitoring for the CET and refund timelocks anymore. This means that
such a CFD would be stuck.

To fix it we have separated the transaction data that becomes
available as events are processed from the logic of whether a
transaction (or a timelock) should be monitored or not. The data
should always be remembered, and only the flags that control the
monitoring logic should change as new events are processed.

[1]: https://github.com/itchysats/itchysats/issues/3028.