fix(monitor): Publish commit transaction on restart
We were never actually setting the `broadcast_commit` (the old
`commit_tx`) field to `Some`, so the commit transaction could never be
re-broadcast on a restart of the application.

The event `ManualCommit` is the only one that warrants the publication
of the commit transaction. Once we process a `CommitConfirmed` event
too, the transaction will not need to be re-broadcast, so we set
`broadcast_commit` to `None`.