chore: Use a dummy CI job to signal overall CI status to bors
It is extremely fragile to need to specify by hand the expanded matrix jobs from
GitHub CI. Furthermore, in case if I job gets skipped, it is not expanded,
making the bors timeout.

Follow suggestion found in bors documentation and set a job that's sole purpose
is to tell bors whether we passed the CI.

See: https://bors.tech/documentation/getting-started/

Note: As a side effect, we won't need to manually update bors.toml whenever we
add a new crate to the workspace.