chore: Reorganise crates in the workspace into crates directory
This makes it simpler to add/remove crate from the workspace (no need to
maintain a list), and also makes it much easier to grok the itchysats repo,
which has many non-crate things too.

This also will allow us to easy match on changed Rust files for the CI
purposes (optimising the CI runs for when we *do not* change any Rust code).