Fix the memory growth by self-implemented batching
This is an attempt to fix the memory growth caused by `rust-electrum-client`'s `batch_script_get_history` implementation.
We chunk up the scripts to be checked into batches and call `script_get_history` for the batches in parallel in separate worker threads.

The results are communicated back using a channel, and are processed as they come; we don't have to rely on any order here.