feat(rollover): Remember state if maker fails to send revocation sk
This ensures that the taker can _always_ safely non-collaboratively
settle the CFD.

Before these changes, the taker could fall behind in the state of the
DLC if the maker failed to send their revocation sk in the last step
of the rollover protocol. The taker would throw away everything
exchanged during that instance of the rollover protocol. This would
mean that the maker would know how to punish the publication of the
commit transaction the taker considered current.

With these changes, the taker remembers the incomplete rollover
details, updating the state of the DLC so that they can safely publish
their latest commit transaction if they want to. The taker also notes
that the maker did not share their revocation sk. This allows the
taker to still insist on rolling over from the previous,
yet-to-be-revoked-by-the-maker commit transaction, like they have been
doing since the introduction of the rollover retry mechanism with
patch e0e643e44d9fa5f03422f4430ffc860d86ef2de6.