Process script updates one by one
Previously we assembled the complete history of all batches and then processed all updates.
With this change we now process them one by one, update the state and only accumulate the events that are ready.

The state update is cheap and can be done as part of processing the request.
Handling the CFD update could be done outside of sync, but we could regard that update as part of sync.
Only once the CFD is updated the sync is actually complete, because only then the state is communicated in the system, so it might be better to keep that within this scope.