Configure timeout for processing responses
If we don't get a response on the receiver for `10` seconds we stop processing responses.
ew expect responses to arrive continuously because we process the requests in parallel.

Upon a timeout we print the amount of responses received (script histories received) and the requests sent (scripts processed).
Note that we still process the histories received knowing that we might not have all responses; it is still valid to update the script history for what we have receive.

Note that it *can* theoretically happen that a sender sends the responses to the receiver but the receiver is gone, but only if for some reason we are stuck while processing. We can't block indefinitely so this is a way forward for now. If we keep seeing a large amount of timeouts we will have to investigate further otherwise this solution is resilient enough for now, because we trigger sync frequently and the next sync should deliver the complete responses.